requirements.txt
fastapi~=0.110.1
psutil~=5.9.8
kafka-python~=2.0.2